Add draft document CrossCompile.txt
This is just a note for people who build flex for a cross-toolchain.
I mention at least how to disable libfl in the first stage and
configure and build libfl for the target in second stage. The
procedures are tricky for now.

I release this draft document in the public domain.
Signed-off-by: Kang-Che Sung <explorer09 @ gmail.com>